Re-use components from course instance tableFields file
Refactor NonClassMeetings table to import the meeting layout components
from the course instances table.

In future, both the non-class meetings table and the course instances
table should be refactored to generalize the components used within.
This work will be done in #362, but for the time being the non-class
meetings table has just been changed to import the components from the
course instances directory