Create and pass down updateCourseHandler
Since our local list of courses lives at the CoursesPage level, we need
to define our update function there and pass it down throw the
CourseInstanceTable to the MeetingModal.

The updateLocalCourse method is intended to handle any update to a
Course; it simply finds the entry with the same id in the list and
replaces it with whatever is passed into the method. This will allow us
to use the same method to handle any other updates, e.g. the instructors
list or the Notes field.