Refactor Body wrapper into CSS grid
The display inside the modal was starting to run up against the limits
of flexbox as more and more 2-dimensional elements were being placed.
This switches to css grid and defines separate areas for the different
elements of the modal, and assigns spacing on a more automated basis.

This also locks the height and width to 75vh/vw; Otherwise, the changes
in content were making the modal shrink/grow in both directions on each
interaction, which makes for a frustrating user experience.