Expose the underlying error to the user
The AxiosError only shows a generic error message in the top level
message field, so we need to dig down into the response.data to access
the more explanatory error message thrown from the server. These
messages still aren't the most human friendly (i.e., an EntityNotFound
error includes the UUID for the element), but should in theory give the
user enought to know to contact us for help.