When state value isChanged is truthy, add a warning message when user exits the modal
The isChanged value indicates whether there has been a change in the meeting modal. The value defaults to false and changes to true when a meeting has been edited, added, or deleted. If the user exits the modal by refreshing the page or closing the modal, a warning message is shown. If the user clicks OK, the modal will close, and if the user clicks cancel, the user can continue making edits in the modal.