Show edit button for courses without meetings
This is slightly outside the scope of this ticket, but showing the
button when a course instance has no meetings fixes an edge case where
the client code attempts to focus the "Edit Meetings" button after
removing all the meetings from a course, throwing an error. It also
helps with test cases where we use that button to get a reference to the
meetings cell in the table.