Add potential room conflict
This adds one additional dummy course AM 22A, that is held on
MON/WED/FRI from 10:30 - 11:45 in Maxwell Dworkin 220. This is the same
room and time as our fake AM 205, but AM 205 is held on TUE/THU. This is
so that we can more easily create a booking conflict in our end-to-end
tests.