Add end-to-end test for editing meetings
When I went to add additional tests for how the server errors are being
handled, I realized that my original "Update Meeting" test was actually
a verbatim duplicate of the "Add Meeting" test. So it passed, but it
wasn't actually testing what it was supposed to. Whoops.

In the process of fixing that test and adding a test of the fail state,
I realized that my original process of trying to find an appropriate
room based on certain criteria just wasn't going to work. Or would have
required waaay too much setup. Instead, we're just going to use AM 205
and AM 22A, which we know can easily be made to conflict (because we're
hard coding that data). This cutsu out some extra logic, and makes for
some simpler assertions to boot.