Update end to end tests to consider addition of window.confirm alert
These tests were updated to consider two contexts: when the modal is saved vs. when it is not saved. The clicking on the "save" button used to be in the beforeEach of each describe block, but they were moved down to each it block of the contexts in which the modal is saved so that the two contexts could be placed on the same level. For each of the scenarios outlined in the describe blocks, we are checking that if changes are made that a window.confirm alert appears.