Extract find() method code into private method
This allows the majority of the code and logic to be re-used, but the
parts specific to each application (i,e: the findMany and fineOne
methods) to be simply chained on in different public methods.

This help to ensure that the findOne and findMany methods return data
that is the same shape (for consistency).