Revert "Extract find() method code into private method"
The initial intent was to allow code-reuse between the "GET ALL" and "GET ONE"
type of service methods. The only problem is that the structure of the
underlying data doesn't lend itself very well to this, and (in the case of the
"GET ONE" type function, would require lots of fancy footwork and fumbling with
data to make it work in a sane manner).

For now, it makes more sense to simply create a record, then re-request the
client side application data