Don't allow non-meridian timestamps
The fromDisplay method is meant to create a new PGTime instance from a
"display time", which here means a timestamp in AM/PM format. After
discussion in code review, it seems weird for it to accept timestamps
that don't have a meridian component, where we just assume that it's
AM. This new implementation is stricter, but should stay closer to the
intent of the code.