Create Request/Response DTOs for instructors
This defines a pair of minimal DTOs, on the assumption that the request
to assign instructors will be made with the same data included in the
CourseInstanceResponseDTO (and rendered on the Courses Page). The
response can then be merged back into the Course page when done.

Since this mirrors the instructors field of the
CourseInstanceResponseDTO, we can also use it as the type for
instructors there.