Assign instructors to courseInstance
This is a little more complicated than it initially seems. We
unfortunately can't just create a bunch of facultyCourseInstance
entities and save them in the database directly, because that would only
add new instructors to the course, not replace/remove them. So we
instead need to create the facultyCourseInstances and then use those to
update the courseInstance, which lets us take advantage of the
orphanedRowAction and cascade options to propagate the changes.