Update injected dependencies
We're adding a few more repositories to the courseInstance service, so
those all need to be added in the module and then included in our tests.
This was also creating some errors in the test output because our API
endpoint requires authentication/authorization, so I needed to properly
register the AuthModule and tweak some of the config settings to get
everything playing nicely.