Set MOCHA_NEEDS_DATABASE variable
This gets/sets an environmnet variable to flag when tests will or won't
need the database container running. In brief:

1. Running `npm run test` alone will set it to 1
2. Running `npm run test -- --integration` will set it to 1
3. Running `npm run test -- --client` will set it to 0
4. Running `npm run test -- --server` will set it to 0

You can also set the variable explicitly from the command line:

1. `MOCHA_NEEDS_DATABASE=0 npm run test` will set it to 0
2. `MOCHA_NEEDS_DATABASE=1 npm run test` will set it to 1
3. `MOCHA_NEEDS_DATABASE=0 npm run test -- --integration` will set it to 0
4. `MOCHA_NEEDS_DATABASE=1 npm run test -- --integration` will set it
   to 1

Running with `--client` or `--server` will always set the variable to 0,
in order to discourage the use of database connection in unit tests

Cases 1 and 3 above will, by default, result in failures as the database
is necessary for most of the integration tests. The  use case for that
format is if you want to use a `.only` to run a specific suite of tests
and don't want to start up the database each time.