Start and stop one database for tests
Based on the value of MOCHA_NEEDS_DATABASE, we'll start and stop a
database at the global level, sharing the container across all the test
suites. This makes the whole test suite run faster, and should also help
avoid errors during the setup and teardown of test suites. The flipside
is that it makes our data more susceptible to cross-contamination
between test suites as we're potentially not starting with a totally
fresh database in each test.

One major change from our previous setup is that we'll now be assigning
a unique name to the test container on each run, and used a somewhat
randomized port. This means that we can potentially run tests in
parallel, and it should avoid errors where an old container blocks the
test suite from running. The downside is that, in cases where the
container fails to stop properly, you could end up with a bunch of
leftover postgres instances hanging around and slowly consuming system
resources. There's not a great solution to that, beyond running `docker
ps` regularly when you're running a lot of tests.

This way of setting things up with global fixtures can cause problems
when used with other mocha features, mainly with running tests in
parallel. Mocha v8.0.0 did add a concept called Root Hooks[1] that may
be better suited to this use case, but we're not on that version yet and
I haven't had a chance to evaluate any other changes.

[1]: https://mochajs.org/#root-hook-plugins