Replace drop methods with clearDatabase
While investigating some of the other features in TypeORM, I found that
there's a `clearDatabase` method that will drop all of the views,
tables, and enums, with CASCADE, in a single transaction, which seems to
resolve our deadlock issues in my local testing. I've still seen some
timeout errors crop up, but those seem to be in frontend code -- we'll
need further investigation there.

I'm going to leave the individual drop methods in the population
services for now in case we do end up needing them in the future.