Set isChanged back to false before the modal closes
This is to fix an issue in which the user creates a new course (successfully submits the Edit Course form), reopens the course via the edit button in the Course Admin table, and tries to exit out of the modal. At this point, the user sees an unsaved changes warning even though they have just opened the modal without making any changes. To fix this issue, I am setting the isChanged state back to false when the modal.