Fix CreateCourse end to end test by waiting for the data in the table to load
The create course button is rendered before all of the courses are returned from the server. The fake axios adapter is using a single superagent instance. Because of this, the POST and GET requests were most likely being sent at the same time, which was causing the request to time out.