Edit names and labels across Faculty filter fields and Faculty modal fields
Across the faculty filter fields and faculty admin modal form fields, there were intersecting names and labels that were causing issues in the end to end faculty admin modal tests in that using the getByLabelText method was getting a reference to the filter fields when it was intended to get a reference to the faculty admin modal form fields. By making the name and labels unique, we should not have this issue going forward of getting references to the unintended fields.