Remove unnecessary await statements and debugging comments
Since the create button of faculty admin and course admin are rendered before the table contents, we do not have to wait for both the table contents and the create button to render before clicking on the "Create New Faculty" and "Create New Course" buttons. It is enough to just have the await statement for the table contents, since by the time the table contents load, the create button has already been rendered.