Set spacing of Instructor lines
I've used flexbox and a nested `flex-grow` to implement the spacing used
in the mockup, which is roughly:

[DELETE] [INSTRUCTOR NAME] [EMPTY SPACE] [UP] [DOWN]

This could have alternatively used CSS grid with named areas for each
field (using a similar nested `grid-area` style), but I think this
flexbox implementation ends up being a little neater, and since we're
really only concerned with positioning along a single axis is more in
tune with the intended use.

One variation from the mockup is that I've included an Inactive version
of the up/down icons for the first/last entries in the list. Partially
this was for convenience since it saves us needing to adjust spacing
dynamically to keep the up and down arrows aligned vertically, but I
think there is a UX benefit of knowing that there is an option to move
faculty up and down, if there were other slots to which they could be
moved.