Add rule to warn in instances of require-default-props rule
After updating the packages, it appears that eslint is now parsing the files in such a way that when default props are not being specified, an error is thrown. This rule seems to have been around for a while, but errors were not being thrown previously. Instead of fixing these lint errors in this PR, it will be done in a separate ticket, because this is a behavioral change that may cause other errors to appear.  https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/require-default-props.md