Update AppRouter tests
The AppRouter tests were updated to check the behavior when there is no metadata in context. The renderRoute function was updated to accept both user and metadata so that we can pass in an empty metadata object. If there is no metadata loaded yet, the loading spinner should still show, since we are waiting for the metadata to load before removing the spinner in the AppRouter file.