Rename CreateNonClassParentDTO
The intent here is to re-use the DTO between create and update operations. On
the face of it, this may seem like an odd choice given the fact that creating
a new DTO is (kind of) "free". Often - the only difference between a create
and an update operation is the existence of an `id` property. In this case,
that value will come from the URL when a `PUT` request is made to
`resource/:id` rendering the difference moot.

In that instance, having a shared DTO object between the two controller
endpoints helps re-enforce the idea that there is no difference in data shape
that the two endpoints accept.