Replace references to `rawAreaList[0]` with an actual Area entity
Previous items from rawAreaList were used to provide mock return values
for various repository methods. However, this data lacks fields commonly
associated with a real instance of `Area` such as the `id` field.

Instead, it is better to use the `appliedMath` stub value from our test
data which is a full entity instance. Most of this was refactored out in
86ca9acddef9a66a15a0f8b33583a68af058f18d, but a few of instances must've
slipped by