Revert "Rename CreateNonClassParentDTO"
At first having a common DTO between create and update operations seemed like a
 good idea as it would help to promote the idea that the two data structures
were identical.

However, the biggest issue with this is that it would mess up the validation,
as some fields are required for a create operation (for example, `area` or
`title`). Whilst, for an update operation _all_ fields are optional so that
unchanged fields aren't updated.

So, whilst the object structures _are_ the same and it _would_ be nice to
share DTOs like that - unfortunately we can't.