Update mark-one version to incorporate role and aria-live properties in ModalMessage component
A course instance end to end test was failing because it could not locate an element with the role of alert in the case that the user provides invalid meeting times and tries to save. The reason for this is because the ModalMessage component did not incorporate the role and aria-live properties until version 0.3.4, and before the error message was wrapped in a div that included these two properties.