Add /faculty/instructors API route
I considered adding this route under the course-instances api namespace
instead, but ultimately decided it belonged here since it's using the
Faculty{Repository,Service} to get the data. I did also initially
implement this with the FacultyListingView since we already defined
displayName there, but that wound up being a complicated query because
it contains duplicates (from the join with facultyCourseInstance), so
any benefit of having displayName defined was kind of nullified.

I've kept the new integration tests in line with the existing ones,
though there's an existing ticket (#415) to refactor some of these so
that all of the tests are run through the API with a mockDB backend. We
should try to update all of those when time permits.