Set a min-height for the modal
Without this height set, the Combobox would appear at the bottom of the
ModalBody area, and the user would need to scroll to see the full list
of option. An alternative would have been to change the overflow-y
setting on the ModalBody, but that would potentially result in the size
of the modal changing as the user types, which seems like a bad UX
tradeoff.