Check for changes before closing
I'd like to explore converting this into a custom hook at some point in
the future, since almost all of the functionality is implemented using
existing React hooks. Then we could replace all of this code with
something like:

```
const { markAsChanged, confirmDiscard } = useConfirmDiscard();
```

Where we'd then call `markAsChanged()` any time we update the data, and
call `confirmDiscard(closeModal)` whenever we'd be calling the
closeHandler without saving.

This would cut down on the amount of copy-pasting, and make the overall
functionality easier to test.

I haven't added any tests for this functionality yet, as I plan to build
out a proper e2e test run when more of the functionality is complete.