Test that assigned instructors don't appear
This is just making sure that we're removing any faculty already
assigned to the given course instance.

As a sidenote, the @testing-library/user-event library has a much
cleaner interface for testing text input events like this. I tried
installing it, but adding the latest version also updated the version of
@testing-library/react, which removes the `wait` and `waitForElement`
methods, and fixing those would have greatly expanded the scope of the
PR so I opted to leave that for another time.