Make onSuccess and onClose props required and provide defaultProps for currentFaculty and currentCourse
The onSuccess and onClose props should always an associated action. They were originally marked as optional when we were first creating the skeletons of these modals. When currentCourse and currentFaculty are null, this means that we are creating a course and faculty rather than editing an existing course or faculty member.