Refactor test suites to use individual test modules
Previously, all tests were sharing one test module. Normally this would
be okay, except that the GET and POST tests stubbed out the database
layer.

Unfortunately, overwriting the test module at the top layer with another
test module caused race conditions where tests would/would not pass
based on whether they were run in isolation or not.

In reality, we probably want to include the database layer in the tests,
so - rather than refactor all of those tests for now - it made more
sense to bring up another test module without the database stubbed out.
This made way for writing tests for the PUT method which _did_ use the
entire database in the tests

The solution ended up being to have test modules in each test suite for
each method. Later, we can go back and refactor so that all the tests
involve the database