Change phone number test data
The DTOs use libphonenumber-js to validate phone numbers. Not only does
libphonenumber validate the format of a phone number
(i.e: (xxx) xxx-xxxx) - it also validates that the actual digits are
valid.

That is to say, that the previous value with a prefix of 128 was
considered NOT valid despite being in the correct format, because 128 is
not a valid US phone number prefix.