Save updated instructor list
This sends a put request to the server to update the list of instructors
associated with the course instance, and passes the resulting list back
up to the onSave method where it will be merged into the local list of
course instances.

I've added some very simple unit tests at present, and will be
implementing fuller end-to-end tests next.