Ensure that the requested field is changed to the correct value
Previously, the assertion just tested that the field had changed, but
not that it had changed to the correct value. Therefore, it changing to
`null` or `undefined` would actually be acceptable under the previous
implementation.

In order to prevent this, it made more sense to test that it had changed
to the correct value.