Change placeholder text while data loads
This is a workaround for an issue that came up in the end-to-end tests,
where some of the PUT requests would fail because they were being
triggered before the GET request for the list of all instructors had
completed. (I'm still not totally clear WHY that's an issue, but I
suspect it's related to something in either supertest or axios). As an
easy way to check that the request is done, we're just changing the
placeholder text on the input field while the GET request is in
progress.