Block the submission of null values
The `IsOptional()` decorator allows `null` values to be submitted as it
prevents any other validators from being run on this field.

These `null` values then raise a 500 error from the server as they
violate the `NOT NULL` column constraint in the database.

The solution was to remove the `IsOptional` field and instead use
`IsNotEmpty` which tests that the value is not equal to empty string,
`null` or `undefined`. Finally, a length requirement of 1 was added to
make sure that an actual value of _some sort_ is submitted (even if it
is just one letter).