Replace render from @testing-library/react with the customRender from test-utils
The tests in Message.test.tsx were failing with the type error: "Cannot read property 'background' of undefined." The reason for this is because the render that was being used does not have the theme context making it so that we could not access "theme.background.color" of the GlobalMessage component. The custom render function of test-utils does have the theme context, so we used this instead.