Add a renderWithMessaging component
The current customRender does not include the message component. There's existing code passing in dispatchMessage to check the call count, so instead of trying to edit the current customRender, I created a new function called renderWithMessaging. Also, customRender cannot be changed to call useReducer, because the hook must be used in a function component. With the new render function, renderWithMessaging, you can pass in the parts of the message needed to check that the actual message is rendered (queue and currentMessage).