Add meetings column to default view
The lack of a "meetings" column in the default view meant that the
meeting modal integration tests were failing as they were expecting an
"Edit Meeting" button to be visible. Since the meetings column was not
visible, this was not the case and the tests thusly failed.

This commit adds the meetings column into the default view to fix these
breaking tests