Add fix for inconsistent number of hooks error
The error that I was getting, "rendered fewer hooks than expected," seems to be coming from the formatInstructors function. When filtering the table, the hook in formatInstructors is not being called, which seems to be causing the inconsistent number of hooks error. This workaround passes in the courseList into filteredCourses to circumvent the issue.