Fix existing test that checks the "Pre", "Study", and "Actual" columns
Now that we have added the area and is SEAS dropdown filters, they are appearing on the third row alongside the "Pre", "Study", and "Actual" column headers. This test was updated to adjust for this change and makes sure that each of the columns show up the correct number of times instead.