Memoize Cell rendering functions
This reduces some of the overhead of rerendering the rows/cells on each
change to the table by checking whether the data has changed before
rendering the content. Checking all those memoizations can still take
time, so there's still a delay before modals open and close. To address
those, I think we'll really need to reduce the number of things being
rendered, likely by implementing some kind of windowing.

I originally tried to implement this with a `useMemo` call around the
entire row rendering logic block, but that ran into one of the drawbacks
to this approach: you cannot call other hooks inside of a `useMemo`
function. This "memoization boundary" might need to be adjusted in the
future, if we need to start adding additional logic in these functions.