Refactor tests in light of hooks
The previous tests were directly calling the retrieveValue or getField
functions, which was violating the Rules of Hooks, i.e., not being
called in a component. This refactors the hook tests so that the
rendering is being done inside of a dummy "TestComponent", and we then
use the react-testing-library methods to make sure the returned value
matches what it should be.