Move logic of rendering up to the CoursesPage table
This commit the course instance table so that the courses page is handling the logic of rendering instead of having the logic in the table. The courses page then passes down the elements that should be displayed in the table, and this represents the filtered data. The filter dropdowns are generated in tableFields via the optional getFilter method.