Use default view columns as initial state
Using an empty array as initial state and then populating it with
columns from the default view was breaking the tests, because the tests
would run too quickly and would have completed before the effect hook
had chance to run and populate the state.

Instead, we can just take the columns from the default view and populate
the initial state with them directly, avoiding the issue entirely.