Block the change handler for mandatory checkboxes
The `fireEvent.click` function doesn't actually click on checkboxes,
instead it simply calls the `onClick` handler. This is a subtle but
important difference, as by doing this, a `disabled` checkbox can still
actually cause changes to occur in the app tests, even though this
wouldn't happen under real world conditions when the app is running.

This is because the `disabled` attribute prevent the checkbox from being
clicked on, but does nothing to prevent the click handler from being
called directly