Generalize the Dropdown component
This sets up the dropdown component so that it can create each filter dropdown by passing in the filter option fields instead of us creating an individual Dropdown for each filter ourselves. This is helping us cut down on code. Also, the label value of the dropdown depends on whether there is a subField (e.g. for the Fall or Spring 'offered' filter).