Add filter field for catalogNumber
A filter field was added for catalogNumber by adding the getFilter property (with a call to generateTextField for the catalogNumber) to the catalogNumber column. The reason for the separation between dropdown filters and text filters in the CoursesPage is so that we can specify a value of false to the property "exact" in the listFilter call so that the text search does not have to be an exact match to yield results. With the dropdown results, we want the "exact" value to be true so that for instance, if you search for the area "CS," you only get "CS" results as opposed to also getting results for "ACS."