Refactor getValue functions into components
This is fundamentally the same thing as a we were doing before, but
using a more familiar Functional Component style in place of the
higher-order function style that was in use previously. However, the
benefit is that we can wrap the functional components output in
React.memo so that it will only re-render if the props change. (It also
arguably makes the implementation code in CourseInstanceTable a little
more clear).