Stabilize buttonRef prop
Passing the buttonRef and buttonId down the component tree separately
was triggering some additional re-renders, as the buttonId would be
flagged as an updated prop whenever it changed, leading to the whole
component tree re-rendering.

This instead moves the id comparison up to the top-level courses page in
a function wrapped in useCallback() -- This gives us a stable function
prop (i.e. it doesn't get flagged as a change on subsequent renders)
that we can use to determine whether the value of buttonRef should be
our real ref or null.