Filter columns in effect hook
Calling the filter() method inside the prop value was triggering
re-renders, as it was returning a new array on each render. This
moves the "tableData" values into a state field, then runs the actual
filtering inside of a useEffect hook that runs whenever the the list of
columns is changed.