Wrap updater functions in useCallback
Defining these directly in the body of the parent function was causing
them to be reassigned on each render of the parent, which can trigger
more re-rendering down the tree whenever they're passed into another
component as a prop. Wrapping in useCallback makes them stable functions
(i.e. they don't change on each render, unless one of the dependencies
in the second-parameter array is changed), so that they won't trigger
re-renders.