Keep modals mounted when invisible
The previous method of conditionally mounting/unmounting the modals when
the visible value in state changed was problematic as it led to the
"children" props of the CoursesPage changing even when the data itself
hadn't changed.

However, keeping the modals mounted required refactoring some of the
data maniupulation, since the component was previously destructuring
values like semester and currentCourse, which could be null. So those
destructured fields have been moved into state values, which are then
being set inside of a useEffect hook.

Along the way I also made the headings of the Modals more consistent
(e.g. using commas instead of hyphens), and moved the template string
that joins the values higher up into the code.

The MeetingModal was also using the formatFacultyNotes function inside
of a prop, which was throwing errors whenever the course/semster values
were null. As a basic solution I moved the prop into a function, which
is instead called inside the component. This fixes the error, but it
might be worth moving that function up into a useCallback hook if it
leads to performance issues.