Specify re-rendering condtions for the table
Adds a memoization wrapper around the entire CourseInstanceTable, and
uses the optional second argument to fine tune when the table should
actuall be re-rendered. Returning TRUE from the functino SKIPS a render,
so we're setting it to only render if the "tableData", "filters" or
"courseList" props changes. In my validation testing this maintained all
the existing functionality, but skips an extra render when opening the
modal, significantly cutting down the overall render time.

It's possible that as we add more filters, modal, and other features
that we'll need to add additional conditions to the function to get the
behavior to work correctly.