Move messageReducer into the Message component
This was kind of a tricky one. Having the reducer in the top-level of
App makes a lot of sense in terms of passing around the state but has
the nasty side effect of causing the entire App tree to re-render every
time that state changes. In practice, this means clicking the "Clear"
button on a message freezes the app for ~2 seconds. Not Good!

So instead we have this convoluted solution that avoids changing the
state at top level by:

1. Creating an empty Ref in App.tsx and passing that into the Message
   component (Refs are mutable, but are treated as stable values so they
   don't trigger a re-render when you update them)
2. Inside of Message, initalize our Reducer and assign Ref.current to
   point at our dispatch function
3. Pass the value of ref.current down through the MessageContext, which
   will allow any child components to update the state inside the
   Message.

The end result is that messages appear and clear quickly, without
needing to re-render any child components except the message itself.
Neat!

This did require a few other small changes:

1. The conditional check for currentMessage is moved inside the Message
   itself so that the children prop of App will always be stable.
2. We need to check for Ref.current before we call dispatchMessage in
   the useEffect hook.
3. We need an extra circuit-breaker in the MessageReducer itself that
   bails out if we don't have any action.

All of this also required some tweaks to our custom render function, the
trickiest of which was the renderWithMessaging function. In our actual
code, we wait to render any actual UI until we have the user data, and
wait to fetch the user data until we acutally have the dispatchMessage
function available. To emulate that in our test function, we're adding
an extra state value, "loading", that's only set to true after
dispatchMessage is there. We can then use that to make sure that the
dispatchMessage is available before we try to render any components
that rely on it.

I wound up essentially rewriting the tests for Message as well. Since
more of the implementation was moved inside the component, we can
approach it as more of a black box test and make sure that calling the
dispatchMessage function and clicking Clear/Next button do what we
expect them to.

Lastly, I changed "clear" to "Clear", to be more consistent with "Next."