Re-implement buttonRef handling
This is sort of a revision of 445e0847. I'm leaving it as a new commit
to help explain the process (also I expect it would be a pain to rebase
back on there).

To step back, all of the changes in this branch have been trying to
improve the performance of the Courses page. From a React perspective,
that mainly means cutting down on unnecessary re-renders, and the main
way to accomplish that is by avoiding changing props or state. The
previous change attempted to do that making a function that returned the
buttonRef, then wrapping it in useCallback so that the function would
only be treated as a new prop if the "buttonModalId" changes (i.e. if a
button was clicked). However, that still meant that clicking the button
would trigger a re-render of the whole CourseInstanceTable. But in
9e4f935d we restricted the circumstances where the table would re-render
and ignored changes to that buttonRef function.

Those two changes wound up being in conflict with each other: Not
re-rendering the table meant the Ref wouldn't actually change, so we
could have fast rendering of the modal, or we could have focus return to
the button on close, but not both. Which is no good.

So I did a little more digging into how Refs works, and came to this
solution, which combines React's [Callback Ref pattern][1] with [a
suggestion from StackOverflow][2]. The idea is that instead of creating
one ref and dynamically assigning it to the button that gets clicked,
we're going to assign a ref to _every_ button, then dynamically select
the ref we want to focus when we call the onClose handler.

Because those buttons are going to be stable between re-renders, we
shouldn't need to recreate the whole list of refs between renders; the
way to do that, naturally, is to store the list of refs inside another
ref! This is taking advantage of the same ideas from 20956961, namely
that the refs created by useRef can really point at _anything_ -- it
just gives us a stable "box" that won't be reassigned between renders.
Combining that with the useCallback hook around our setButtonRef
function means that we have a stable value that will persist between
renders.

One additional wrinkle is that the value we're using as our key isn't
derivable from just the HTML node, so we need to fall back to my old
favorite, the higher-order function. We're really passing down a
function we can call with unique modalButtonId value, which will then
create our CallbackRef that adds the ref to our table.

Finally, I just want to amplify the ["Caveats with callback refs"][3]
from the react docs (which I overlooked on first pass). Essentially, the
callback passed to the ref value will be called twice, first with null
then with the actual element. So before we add it to our refTable, we
need to make sure that the node actually exists.

[1]: https://reactjs.org/docs/refs-and-the-dom.html#callback-refs
[2]: https://stackoverflow.com/questions/67698802/conditionally-assign-ref-in-react
[3]: https://reactjs.org/docs/refs-and-the-dom.html#caveats-with-callback-refs