Move the filtering values of filterByInstructors into utils file and add tests to check
To make the code more modular, the filtering of instructors was moved to a utils directory. The tests were updated to differentiate between the utils function that filters multi year plans from this one, which filters courses by instructors. Tests to check the behavior of the fall and spring instructor filters were also added.