Debounce the setting of filters state
When the user types into the filter field, it wait 100 milliseconds to check if the user is still typing. This is so that the filters state is not necessarily updated every single time the user types a letter. If the user types an additional letter within the 100 ms, the updating of the state is delayed. This helps us decrease the data processing and hopefully helps with decreasing the observed wait time by the user.