Memoize the Course Instance Table Body
Previously, whenever the user typed into the filter field, this would cause the entire table to re-render in order to show the user what they typed in the field. This is because the table header (which contains the filters) and table body were both in 1 component, the CourseInstanceTable. By separating out the body of the table, the user can quickly see what they type in the text filter field without having to re-render the entire table contents in between key strokes.