Fix TypeORM Error: Driver Not Connected error
The "driver not connected" error was showing up previous to scoping the fake timer using this "toFake" property. Worked with R Main on this issue - it seems that you need to scope the timer to directly what you will be using it for to prevent it from causing ripple effects in other unrelated tests.