Update instructor and meeting modal onSave behavior
Originally, both the instructor and meeting modals would close immediately after successfully saving changes. The table body would temporarily display outdated data and then render the updated data a second or so later when the new set of courses finished setting in the useEffect block.

As a result of this update, there are two new states, shouldCloseMeetingModal and shouldCloseInstructorModal, which indicate that the user has clicked "save" in one of these modals. These two states allow us to be able to close the instructor and meeting modals only after the courses instances have been set. This means that these modals do not close until the underlying course instance table has updated with the updated table data.